Updated if statements
Problem with previous version was if there was a space on the right and in front then the regular wall follower method would run. This would mean the robot would move off to the right hand side to equal the distances between the walls. Therefore a forwards method is required to skip past the open right wall first.